pyproject.toml
[tool.poetry]
name = "test_python"
version = "0.1.0"
description = ""
authors = ["Rodolfo De Nadai <rdenadai@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<=3.10"
httpx = "^0.21.1"
types-setuptools = "^57.4.2"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.910"
bandit = "^1.7.1"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
